# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Public REST API for agentserver. Generated from\nGo handler annotations via swaggo/swag; see\ndocs/api/README.md for regeneration instructions.",
    "title": "agentserver API",
    "contact": {},
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/tasks/poll": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Poll for pending tasks (proxy_token auth)",
        "parameters": [
          {
            "description": "Sandbox ID (defaults to token's sandbox)",
            "name": "sandbox_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentTaskPollItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/tasks/{id}": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get a task by ID (proxy_token auth)",
        "parameters": [
          {
            "description": "Task ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskDetail"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/tasks/{id}/status": {
      "put": {
        "tags": [
          "Agent"
        ],
        "summary": "Update task status (proxy_token auth)",
        "parameters": [
          {
            "description": "Task ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTaskStatusRequest"
              }
            }
          },
          "description": "Status update",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{id}": {
      "get": {
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "tags": [
          "Agent"
        ],
        "summary": "Get a task by ID",
        "parameters": [
          {
            "description": "Task ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Include output events",
            "name": "include_output",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskDetail"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/cancel": {
      "post": {
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "tags": [
          "Agent"
        ],
        "summary": "Cancel a task",
        "parameters": [
          {
            "description": "Task ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskCancelResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "task already finished",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{wid}/tasks": {
      "get": {
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "tags": [
          "Agent"
        ],
        "summary": "List delegated tasks for a workspace",
        "parameters": [
          {
            "description": "Workspace ID",
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentTaskItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "not a member",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "tags": [
          "Agent"
        ],
        "summary": "Create a delegated task in a workspace",
        "parameters": [
          {
            "description": "Workspace ID",
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentTaskCreateRequest"
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "target agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "requestBodies": {
...
      "AgentTaskCreateRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentTaskCreateRequest"
            }
          }
        },
        "description": "Task details",
        "required": true
      }
    },
    "securitySchemes": {
      "CookieAuth": {
        "description": "Session cookie set by POST /api/auth/login. All non-auth endpoints assume this cookie.",
        "type": "apiKey",
        "name": "agentserver-token",
        "in": "cookie"
      }
    },
    "schemas": {
...
      "AgentTaskCancelResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "cancelled"
          }
        }
      },
      "AgentTaskCreateRequest": {
        "type": "object",
        "required": [
          "prompt",
          "target_id"
        ],
        "properties": {
          "delegation_chain": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max_budget_usd": {
            "type": "number",
            "example": 1
          },
          "max_turns": {
            "type": "integer",
            "example": 50
          },
          "prompt": {
            "type": "string",
            "example": "Run the test suite"
          },
          "requester_id": {
            "type": "string"
          },
          "skill": {
            "type": "string",
            "example": "testing"
          },
          "system_context": {
            "type": "string"
          },
          "target_id": {
            "type": "string",
            "example": "sandbox-uuid"
          },
          "timeout_seconds": {
            "type": "integer",
            "example": 300
          }
        }
      },
      "AgentTaskCreateResponse": {
        "type": "object",
        "required": [
          "status",
          "task_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "task_id": {
            "type": "string"
          }
        }
      },
      "AgentTaskDetail": {
        "type": "object",
        "required": [
          "created_at",
          "prompt",
          "status",
          "target_id",
          "task_id",
          "workspace_id"
        ],
        "properties": {
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "failure_reason": {
            "type": "string",
            "nullable": true
          },
          "num_turns": {
            "type": "integer"
          },
          "prompt": {
            "type": "string"
          },
          "requester_id": {
            "type": "string"
          },
          "result": {
            "nullable": true
          },
          "session_id": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "target_id": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "total_cost_usd": {
            "type": "number",
            "nullable": true
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
      "AgentTaskItem": {
        "type": "object",
        "required": [
          "created_at",
          "prompt",
          "status",
          "target_id",
          "task_id"
        ],
        "properties": {
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "num_turns": {
            "type": "integer"
          },
          "prompt": {
            "type": "string"
          },
          "requester_id": {
            "type": "string"
          },
          "skill": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "target_id": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "total_cost_usd": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "AgentTaskPollItem": {
        "type": "object",
        "required": [
          "prompt",
          "task_id"
        ],
        "properties": {
          "max_budget_usd": {
            "type": "number"
          },
          "max_turns": {
            "type": "integer"
          },
          "prompt": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "system_context": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          }
        }
      },
      "AgentTaskStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "failure_reason": {
            "type": "string"
          },
          "num_turns": {
            "type": "integer"
          },
          "result": {},
          "status": {
            "type": "string",
            "example": "completed"
          },
          "total_cost_usd": {
            "type": "number",
            "nullable": true
          }
        }
      },
...
    }
  }
}